refactor: Consolidate some smaller install tasks into one file
